factored out utils which don't rely on tensornetwork
